[refactor] input 클래스 분리 및 그로 인한 app 수정
